refactor: extract `defaultColor` constant
