+++ packages/ngx-foundation-sites/src/lib/colors/default-color.ts
@@ -1,0 +1,3 @@
+import { FasColor } from './color';
+
+export const defaultColor: FasColor = 'primary';
